==> deliverables/notebooks/senior project/for CSU competition/02 - Cleaning data.ipynb <==
import re
import pandas as pd
import numpy as np
import csv
import json
import os

os.getcwd()

df = pd.read_csv(open("..\\..\\data\\Weekly_data.csv", encoding = "utf-8"))

def isnull(x):
    """
    """
    return x != x

def clean_lyric(lyric, uni_replace):
    
    repl = {"\[.+?\]": "", 
            "\s{2,}": "\n", 
            "lyrics": "", 
            "\(": "", 
            "\)": "",
            "\<u\+0092\>": "'",
            "\<u\+0097\>": "-",
            "\<u\+00..\>": "",
            " '{0,}": " ",
            "in'{1} ": "ing ",
            "\.{0,} ": " ",
            " \.{0,}": " ",
            ' "{0,}': " ",
            '"{0,} ': " ",
            }
    
    for i in repl:
        lyric = re.sub(i, repl[i], lyric, flags = re.I | re.S)
        
    for i in uni_replace:
        lyric = re.sub(i, "", lyric, flags = re.I | re.S)
        
    
    
    return lyric

filler = [" and ", " featuring ", " & ", " x ", " / "]
repl = {"(\"|”).+?(\"|”)": "", "\s+": " ", "\s,": ","}
textcols = ["Artists", "Name", "Genre", "Writing Credits", "Lyrics"]
unicode_replace  = json.load(open("..//..//data//unicoderemove.json", "r"))

for i in textcols:
    for j in range(len(df)):
        if not isnull(df.loc[j, i]):
            if i == "Artists":
                for k in filler:
                    df.loc[j, i] = re.sub(k, ", ", df.loc[j, i], flags = re.I | re.S)
            elif i == "Lyrics":
                df.loc[j, i] = clean_lyric(df.loc[j, i], unicode_replace)
            else:
                for k in repl:
                    df.loc[j, i] = re.sub(k, repl[k], df.loc[j, i], flags = re.I | re.S)
                df.loc[j, i] = df.loc[j, i].strip()

# df[df.Name == "No Lie"].reset_index(drop = True).loc[0].Lyrics

## Add songwriters

writers = json.load(open("..//..//data//composers.json", "r"))

composers = []
for i in range(len(df)):
    for j in writers:
        if not isnull(df.loc[i, "Writing Credits"]):
            if re.search(j, df.loc[i, "Writing Credits"], flags = re.I | re.S) is not None:
                if re.search(j, df.loc[i, "Artists"], flags = re.I | re.S) is None:
                    composers.append(True)
                    break
    else:
        composers.append(False)

len(composers)

df["Songwriter"] = np.array(composers)

# we drop NA's here for convenience now....will fix later
df = df[df.Lyrics == df.Lyrics]
df.to_csv("..//..//data//Weekly_data_clean.csv", index = False)

import requests
thingy = str(requests.get("https://raw.githubusercontent.com/dwyl/english-words/master/words.txt").content)

import re
[thingy[x.start() - 10: x.end() + 10] for x in list(re.finditer("shit", str(thingy)))]
